% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_boot.R
\name{CI_scoreBoot}
\alias{CI_scoreBoot}
\title{confidence interval for the estimated genetic model}
\usage{
CI_scoreBoot(data, scoreF_data, boot_Num = NULL,
  method_se = "percentile", alpha_CI = NULL, grid)
}
\arguments{
\item{data}{a data set consists of phenotype, genotype under additive genetic model assumption, genotype expanded with grid interval, and covariate if applicable.}

\item{scoreF_data}{a list of score functions generated by effScore.}

\item{boot_Num}{a numeric value of bootrapping replications.}

\item{method_se}{a method to compute standard error (default='percentile'). Users must choose a method in c('standard', 'percentile', 'BC', 'BCa').
'standard' calculates a standard error of the bootstrap samples. 'percentile' uses percentiles of the bootstrap distribution to define the percentile confidence limits. 'percentile' method corrects for non-normality of the estimator of interest.
'BC' is a Bias-Corrected percentile method which corrects for narrowness bias of percentile CIs. The bias-correction factor isestimated using the inverse cdf of standard normal distribution. 'BCa' is a bias-corrected and accelerated
percentaile method which corrects for narrowness bias and for nonconstant standard error of the estimator of interest (acceleration). The acceleration factor estimates
the rate of change of the standard error of the estimator.}

\item{alpha_CI}{a significance level (default=0.05).}

\item{grid}{an interval of grid search (default=0.01; range: 0 < grid < 1). A lower value of grid gives denser intervals.}
}
\value{
a list of estimated genetic model and its confidence interval.
}
\description{
Confidence interval for the estimated genetic model
}
\author{
Yeonil Kim
}
