% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effScore.R
\name{effScore}
\alias{effScore}
\title{Score functions}
\usage{
effScore(pheno, geno_grid, covr = NULL, y_hat, boot = FALSE,
  boot_Num = NULL)
}
\arguments{
\item{pheno}{a vector of phenotype.}

\item{geno_grid}{a n x m matrix of genotype (n=the number of sample size; m=the number of interval between 0 and 1).}

\item{covr}{a vector of covariate.}

\item{y_hat}{a vector of fitted phenotype values under the null hypothesis.}

\item{boot}{a logical value indicating whether to use bootstrap for confidence interval of the estimated genetic model.}

\item{boot_Num}{a numeric value of bootstrap replications.}
}
\value{
a list of score functions, variances, bootstrap samples of the score functions, and bootstrap samples of the variances.
}
\description{
Score functions
}
\author{
Yeonil Kim
}
