% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSNP.R
\name{singleSNP}
\alias{singleSNP}
\title{Simulating Single SNP Data}
\usage{
singleSNP(size, geno_eff, grid = 0.01, c, MAF, pv = 0.5,
  cov_eff = 0.5)
}
\arguments{
\item{size}{a numeric value of sample size.}

\item{geno_eff}{a numeric value of genotype effect on phenotype.}

\item{grid}{an interval of grid search (default=0.01; range: 0<grid<1). A lower value of grid gives denser intervals.}

\item{c}{a numeric value of heterozygous genotype effect (0<=c<=1).}

\item{MAF}{a numeric value of minor allele frequency (0<MAF<=0.5).}

\item{pv}{a numeric value of prevalence of disease (default=0.5).}

\item{cov_eff}{a numeric value of covariate effect on phenotype (default=0.5).}
}
\value{
a data set which consists of phenotype, genotype under additive assumption of genetic model, genotype under true genetic model, genotype matrix expanded with grid interval, and covariate.
}
\description{
Simulating Single SNP Data
}
\author{
Yeonil Kim
}
